Revert "Send Slack Events API invocations to AWS SNS"
