+++ aws/package.json
@@ -14,7 +14,6 @@
     "@types/common-tags": "^1.8.0",
     "axios": "~0.21.0",
     "common-tags": "^1.8.0",
-    "luxon": "^1.25.0",
     "rss-parser": "^3.9.0",
     "source-map-support": "^0.5.10",
     "tsscmp": "^1.0.6"
@@ -24,7 +23,6 @@
     "@types/aws-lambda": "^8.10.17",
     "@types/jest": "^26.0.15",
     "@types/js-yaml": "^3.12.5",
-    "@types/luxon": "^1.25.0",
     "@types/node": "^12.19.3",
     "@types/serverless": "^1.78.6",
     "@types/tsscmp": "^1.0.0",
@@ -40,7 +38,6 @@
     "serverless": "^2.12.0",
     "serverless-dynamodb-local": "^0.2.39",
     "serverless-offline": "^6.8.0",
-    "serverless-offline-sns": "^0.68.0",
     "serverless-webpack": "^5.2.0",
     "ts-jest": "^26.4.4",
     "ts-loader": "^5.3.3",
