+++ aws/src/slash/invite-all/index.ts
@@ -3,7 +3,7 @@ import { APIGatewayProxyHandlerV2 } from 'aws-lambda';
 import { envvar } from '@/lib/envvar';
 import { slack } from '@/lib/slack/client';
 import { WebAPICallResult } from '@slack/web-api';
-import { verify } from '@/lib/slack/verify';
+import { verify } from '../lib/verify';
 import { SlashParams } from '../lib/params';
 
 // TODO: ライブラリに切り出す
