+++ aws/tsconfig.json
@@ -1,12 +1,12 @@
 {
   "compilerOptions": {
-    "lib": ["es2019"],
+    "lib": ["es2017"],
     "removeComments": true,
     "moduleResolution": "node",
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "sourceMap": true,
-    "target": "es2019",
+    "target": "es2017",
     "outDir": ".build",
     "esModuleInterop": true,
     "types": [
